mockApi for manage user
