+++ cypress.config.ts
@@ -3,6 +3,7 @@ import { resetStubs } from './integration_tests/mockApis/wiremock'
 import auth from './integration_tests/mockApis/auth'
 import tokenVerification from './integration_tests/mockApis/tokenVerification'
 import preview from './integration_tests/mockApis/preview'
+import manageUsers from './integration_tests/mockApis/manageUsers'
 
 export default defineConfig({
   chromeWebSecurity: false,
@@ -21,6 +22,7 @@ export default defineConfig({
       on('task', {
         reset: resetStubs,
         ...auth,
+        ...manageUsers,
         ...tokenVerification,
         ...preview,
       })
